Refactor routes folder to follow interface implementation
