+++ server/index.js
@@ -3,36 +3,25 @@ const http = require('http');
 const cron = require('node-cron');
 const dotenv = require('dotenv');
 
-const MiddlewareManager = require('./middlewares/MiddlewareManager');
+dotenv.config();
 
-// Routes Import
-const authRoutes = require('./routes/auth');
-const githubRoutes = require('./routes/github');
-const paymanRoutes = require('./routes/payman');
-const discordRoutes = require('./routes/discord');
-const businessRoutes = require('./routes/business');
-const contractorRoutes = require('./routes/contractor');
+const MiddlewareManager = require('./middlewares/MiddlewareManager');
+const RouteManager = require('./routes/RouteManager');
 
 // Jobs
 const syncGitHubIssues = require('./jobs/syncGitHubIssues');
-
-dotenv.config();
 
 const app = express();
 const server = http.createServer(app);
 const initSocket = require('./sockets/chat');
 
-// Middlewares
+// Middlewares.
 const middlewareManager = new MiddlewareManager();
 middlewareManager.applyMiddlewares(app);
 
-// Routes
-app.use('/api/auth', authRoutes);
-app.use('/api/payman', paymanRoutes);
-app.use('/api/github', githubRoutes);
-app.use('/api/discord', discordRoutes);
-app.use('/api/business', businessRoutes);
-app.use('/api/contractor', contractorRoutes);
+// Routes.
+const routeManager = new RouteManager();
+routeManager.applyRoutes(app);
 
 // Socket
 initSocket(server);
