--- server/routes/auth.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const { loginUser, signupUser } = require('../controllers/authController');
-
-router.post('/login', loginUser);
-router.post('/signup', signupUser);
-
-module.exports = router;
